+++ src/index.ts
@@ -2,4 +2,5 @@ import PasteClient from "./PasteClient.js";
 
 export { PasteClient };
 export default PasteClient;
-export * from "./types";
+
+export * from "./types.js";